test: blaise update case models for FRS & LMS
refs: BLAIS5-4351
